IDPS controller and agent
This MR is an out-of-box solution to integrate Lastline Suricata with Antrea.

Added entry binary files:
  - antrea-idps-agent, in cmd/idps-agent.
  - antrea-idps-controller, in cmd/idps-controller.

Added docker images:
  - antreainterworking/idps, which includes antrea-idps-agent and antrea-idps-controller.
  - antreainterworking/suricata, based on Lastline Suricata image.

Added manifests and tool files:
  - build/charts/idps, source files to generate idps.yml.
  - build/yamls/idps.yml, the out-of-box yaml file to deploy everything.
  - hack/generate-manifest-idps.sh, the script to generate idps.yml.

The Antrea IDPS Controller (antrea-idps-controller) has the following components:

  - Signature Manager, which is responsible for fetching the latest signature data
    from different signature providers.
  - IDPSPolicy Controller, which is responsible for mirroring an IDPSPolicy object
    to a TrafficControl object.
  - APIServer, which is responsible for exposing the latest signature data fetched
    by Signature Manager via HTTP APIs. Note that, the authentication and authorization
    of the APIServer are delegated to Kubernetes.
  - Registration Controller, which is responsible for validating the registration of
    Antrea to NSX.

Signature Manager is used to manage multiple signature providers. Currently, we
have only one signature provider NTICS. For every signature provider, it syncs the
latest signature data periodically, and it also provides a unified interface which
is used get the signature data for callers. Note that, the signature data is not
stored in local filesystem. It is only cached in memory.

IDPSPolicy is a convenient CRD to select Pods with AppliedTo and apply the selected
Pods to Suricata engine. Its functionality is implemented by IDPSPolicy Controller.
When creating a IDPSPolicy, a corresponding TrafficControl will be also created.

APIServer is responsible for exposing the latest signature data in Signature Manager
via HTTP API. The signature data of every signature provider is exposed by the URI
`/signatures/<signature provider name>`.

Registration Controller is used by signature provider NTICS to verify the registration
to NSX. The controller watches a CRD AntreaRegistration which contains an encrypted
timestamp. The registration is considered valid if the timestamp is decrypted
successfully and the timestamp is not 10 minutes earlier than from the current time.

The Antrea IDPS Agent (antrea-idps-agent) has the following components:

  - Signature Controller, which is responsible for syncing the latest signature data
    from the APIServer in IDPS Controller.
  - Suricata, which is the IDS engine.

Signature Controller is used to watch CRD IDPSSignatureProviderInfo and sync the
signature data from the APIServer in IDPS Controller. After syncing signature data,
Suricata will be reloaded to load the signature data.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>
Co-authored-by: Ran Gu <gran@vmware.com>