Support multiple Egress IPs for AWS multi-AZ failover
A Kubernetes cluster on AWS typically span 3 availability zones and 3
subnets. The egress IP needs to be routable in the underlay network and
usually resides in the same subnet of the Node hosting it. Therefore,
there may be a situation where no available nodes are eligible to host
an egress IP address if all Nodes in a subnet are down, interrupting
egress traffic of the workloads that use the IP.

To support AWS multi-AZ failover, we support configuring multiple Egress
IPs for a single Egress, then users could allocate one Egress IP from
each AZ. At any given moment, only one Egress IP will take effect. When
there is no eligible Node for the effective Egress IP, the next IP will
be promoted as the effective one.

Besides, AWS VPC doesn't forward ARP broadcasts. The Egress IPs must be
configured as the secondary IPs of the Node instances to receive the
response traffic from external. We add a cloud provider interface for
managing the secondary IPs on AWS, and a cloud controller for Egress
which synchronizing the Egress IPs to the secondary IPs of the Node
instances.

Signed-off-by: Quan Tian <qtian@vmware.com>
Co-authored-by: Shuyang Xin <gavinx@vmware.com>
Co-authored-by: Wenqi Qiu <wenqiq@vmware.com>