Fix version parser.
After we introduced the `versions.config` file to support `cayman_antrea_package` target , it will be used as the source 
of truth of various version parameters when the branch name 
parse failed(eg. for topic branches). However, we will also 
need to load this file with the supported branch(eg: `vmware-2.1.0+vmware.1`, `vmware-master`...) because we 
introduced new version variables in this file. So this file 
will always need to be loaded.